changed the structure of datasets

+++ src/utils/misc.py
@@ -18,11 +18,12 @@ def one_hot(targets, cl):
 
 
 def one_vs_all_one_hot(targets, tasks, classes):
-    targets = targets.type(torch.LongTensor).view(-1)
+    targets = targets.view(-1)
     targets_onehot = torch.zeros(targets.size()[0], len(classes))
     for i, t in enumerate(targets):
         if t in tasks:
             targets_onehot[i][t] = 1
+
     return targets_onehot
